refacto: extract drug from pharmacy
